[Multimodal] Reorganize video decoder backends (#49155)
