+++ tests/multimodal/test_gpu_ipc_memory.py
@@ -6,7 +6,6 @@ import time
 
 import pytest
 
-import vllm.config.multimodal as multimodal_config_module
 from vllm.config.multimodal import MultiModalConfig
 from vllm.multimodal.gpu_ipc_memory import (
     MultiModalGPUMemoryPool,
@@ -15,10 +14,10 @@ from vllm.multimodal.gpu_ipc_memory import (
     reserve_mm_ipc_gpu_memory,
     set_mm_gpu_ipc_pool,
 )
-from vllm.multimodal.video import (
+from vllm.multimodal.video_decoders import PYNVVIDEOCODEC_VIDEO_BACKEND
+from vllm.multimodal.video_decoders.pynvvideocodec import (
     PYNVVIDEOCODEC_CUDA_CONTEXT_BYTES,
     PYNVVIDEOCODEC_DECODER_GPU_MEMORY_BYTES,
-    PYNVVIDEOCODEC_VIDEO_BACKEND,
 )
 from vllm.utils.mem_constants import GiB_bytes
 
@@ -186,11 +185,7 @@ def test_reserve_mm_ipc_gpu_memory_raw_frame_budget_only(
     monkeypatch: pytest.MonkeyPatch,
     video_backend: str | None,
 ):
-    monkeypatch.setattr(
-        multimodal_config_module.envs,
-        "VLLM_VIDEO_LOADER_BACKEND",
-        "opencv",
-    )
+    monkeypatch.setenv("VLLM_VIDEO_LOADER_BACKEND", "opencv")
     mm_config = _mm_config(
         mm_ipc_gpu_memory_gb=0.25,
         video_backend=video_backend,
@@ -202,11 +197,7 @@ def test_reserve_mm_ipc_gpu_memory_raw_frame_budget_only(
 def test_reserve_mm_ipc_gpu_memory_includes_pynvvideocodec_decoder_budget(
     monkeypatch: pytest.MonkeyPatch,
 ):
-    monkeypatch.setattr(
-        multimodal_config_module.envs,
-        "VLLM_VIDEO_LOADER_BACKEND",
-        "opencv",
-    )
+    monkeypatch.setenv("VLLM_VIDEO_LOADER_BACKEND", "opencv")
     mm_config = _mm_config(
         mm_ipc_gpu_memory_gb=0.25,
         video_backend=PYNVVIDEOCODEC_VIDEO_BACKEND,
@@ -221,11 +212,7 @@ def test_reserve_mm_ipc_gpu_memory_includes_pynvvideocodec_decoder_budget(
 def test_reserve_mm_ipc_gpu_memory_uses_env_video_backend(
     monkeypatch: pytest.MonkeyPatch,
 ):
-    monkeypatch.setattr(
-        multimodal_config_module.envs,
-        "VLLM_VIDEO_LOADER_BACKEND",
-        PYNVVIDEOCODEC_VIDEO_BACKEND,
-    )
+    monkeypatch.setenv("VLLM_VIDEO_LOADER_BACKEND", PYNVVIDEOCODEC_VIDEO_BACKEND)
     available_bytes = 4 * GiB_bytes
 
     assert reserve_mm_ipc_gpu_memory(available_bytes, _mm_config()) == (
@@ -236,11 +223,7 @@ def test_reserve_mm_ipc_gpu_memory_uses_env_video_backend(
 def test_reserve_mm_ipc_gpu_memory_scales_decoder_budget_by_api_servers(
     monkeypatch: pytest.MonkeyPatch,
 ):
-    monkeypatch.setattr(
-        multimodal_config_module.envs,
-        "VLLM_VIDEO_LOADER_BACKEND",
-        PYNVVIDEOCODEC_VIDEO_BACKEND,
-    )
+    monkeypatch.setenv("VLLM_VIDEO_LOADER_BACKEND", PYNVVIDEOCODEC_VIDEO_BACKEND)
     available_bytes = 8 * GiB_bytes
 
     assert reserve_mm_ipc_gpu_memory(
@@ -253,11 +236,7 @@ def test_reserve_mm_ipc_gpu_memory_scales_decoder_budget_by_api_servers(
 def test_reserve_mm_ipc_gpu_memory_uses_configured_hw_decoders(
     monkeypatch: pytest.MonkeyPatch,
 ):
-    monkeypatch.setattr(
-        multimodal_config_module.envs,
-        "VLLM_VIDEO_LOADER_BACKEND",
-        "opencv",
-    )
+    monkeypatch.setenv("VLLM_VIDEO_LOADER_BACKEND", "opencv")
     available_bytes = 4 * GiB_bytes
     mm_config = _mm_config(
         video_backend=PYNVVIDEOCODEC_VIDEO_BACKEND,
